Create README.md
This model/project aims to achieve improved test set accuracy by implementation of Residual Networks (ResNets) on datasets where training data is limited. An accuarcy of 74% was achieved when Residual Layers were introduced to the model as compared to 68% when the same model was used on the training data without any Residual Layers.